Moved encryption logic out of enigma class
