+++ test/encryption_test.rb
@@ -1,0 +1,11 @@
+require 'minitest/autorun'
+require 'minitest/pride'
+require './lib/encryption'
+
+class EncryptionTest < Minitest::Test
+
+  def test_it_exists
+    encryption = Encryption.new
+    assert_instance_of Encryption, encryption
+  end
+end
